requirements.txt
attrs==21.2.0
click==8.0.3
colorama==0.4.4
dictpath==0.1.3
Flask==2.0.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
more-itertools==8.12.0
openapi-core==0.14.2
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
parse==1.19.0
pyodbc==4.0.32
PyYAML==6.0
six==1.16.0
Werkzeug==2.0.2
